Refactor FormContent into its own component. Add WizardDialog component.
